KRACOEUS-6955 : Common Questionnaire core 1.Move classes 2.Common Questionnaire core autowire QuestionnaireServiceImpl 3.decouple dependency from VersioningServiceImpl
