+++ coeus-code/src/main/java/org/kuali/coeus/common/questionnaire/framework/core/MultiQuestionableFormInterface.java
@@ -13,20 +13,20 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.kuali.kra.questionnaire;
+package org.kuali.coeus.common.questionnaire.framework.core;
 
 /**
  * 
- * This class determines what functions must be present on a form that uses questionnaire.
+ * This class determines what functions must be present on a form that uses questionnaire in more than one use.
+ * Note, all implementing functions must contain the same number of array elements.
  * KraTransactionalFormBase.populateFalseCheckboxes will call these function to determine if a field is
  * associated with a questionnaire answer field.
  */
-public interface QuestionableFormInterface {
-    public static final String DEFAULT_MIDDLE = "].answers[";
-    public static final String DEFAULT_END = "].answer";
+public interface MultiQuestionableFormInterface {
+    public static final String DEFAULT_MIDDLE = QuestionableFormInterface.DEFAULT_MIDDLE;
+    public static final String DEFAULT_END = QuestionableFormInterface.DEFAULT_END;
     
-    public String getQuestionnaireFieldStarter();
-    public String getQuestionnaireFieldMiddle();
-    public String getQuestionnaireFieldEnd();
-    public String getQuestionnaireExpression();
+    public String[] getQuestionnaireFieldStarters();
+    public String[] getQuestionnaireFieldMiddles();
+    public String[] getQuestionnaireFieldEnds();
 }
