+++ coeus-code/src/main/java/org/kuali/coeus/sys/framework/model/KcTransactionalDocumentFormBase.java
@@ -22,8 +22,8 @@ import org.kuali.coeus.sys.framework.validation.SoftError;
 import org.kuali.kra.authorization.KraAuthorizationConstants;
 import org.kuali.kra.infrastructure.KeyConstants;
 import org.kuali.kra.krms.KcKrmsConstants;
-import org.kuali.kra.questionnaire.MultiQuestionableFormInterface;
-import org.kuali.kra.questionnaire.QuestionableFormInterface;
+import org.kuali.coeus.common.questionnaire.framework.core.MultiQuestionableFormInterface;
+import org.kuali.coeus.common.questionnaire.framework.core.QuestionableFormInterface;
 import org.kuali.rice.kim.api.KimConstants;
 import org.kuali.rice.kns.document.authorization.DocumentAuthorizerBase;
 import org.kuali.rice.kns.web.struts.form.KualiTransactionalDocumentFormBase;
