+++ coeus-code/src/main/java/org/kuali/kra/questionnaire/CoeusModuleValuesFinder.java
@@ -16,6 +16,7 @@
 package org.kuali.kra.questionnaire;
 
 import org.kuali.coeus.common.framework.module.CoeusModule;
+import org.kuali.coeus.common.questionnaire.framework.core.QuestionnaireService;
 import org.kuali.coeus.sys.framework.service.KcServiceLocator;
 import org.kuali.rice.core.api.util.ConcreteKeyValue;
 import org.kuali.rice.core.api.util.KeyValue;
