+++ coeus-it/src/test/java/org/kuali/kra/questionnaire/QuestionnaireServiceTest.java
@@ -22,6 +22,8 @@ import org.jmock.lib.concurrent.Synchroniser;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
+import org.kuali.coeus.common.questionnaire.framework.core.QuestionnaireService;
+import org.kuali.coeus.common.questionnaire.impl.core.QuestionnaireServiceImpl;
 import org.kuali.coeus.sys.framework.service.KcServiceLocator;
 import org.kuali.kra.test.infrastructure.KcIntegrationTestBase;
 import org.kuali.rice.krad.UserSession;
